pyproject.toml
[tool.poetry]
name = "type-based-creator"
version = "0.1.0"
description = ""
authors = ["Ghennadi Colievschii <meatraven@gmail.com>"]
readme = "README.md"
packages = [{include = "type_based_creator", from = "apps"}]

[tool.poetry.dependencies]
python = "^3.10.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"

[tool.poetry.group.test.dependencies]
mypy = "^1.3.0"
pylint = "^2.17.4"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:Using the __implements__ inheritance pattern for BaseReporter is no longer supported. Child classes should only inherit BaseReporter:DeprecationWarning",
    "ignore::pytest.PytestRemovedIn8Warning",
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning",
]

[tool.mypy]
show_error_codes = true
disable_error_code = ["import"]

[tool.isort]
profile = "black"